core/requirements.txt
Django==3.2.5
psycopg2-binary
python-decouple
Pillow
djangorestframework
django-cors-headers
